+++ src/file_processor/mod.rs
@@ -1,0 +1,4 @@
+pub mod date;
+pub mod directory;
+pub mod file;
+mod normalization;
